+++ lib/src/enrollments/enrollments_component.dart
@@ -1,8 +1,8 @@
 library plato.cap.components.enrollments;
 
 import 'package:angular/angular.dart';
-import 'package:plato.cap/src/enrollments/enrollment.dart';
 
+import 'package:plato.cap/src/enrollments/enrollment.dart';
 import 'package:plato.cap/src/enrollments/enrollments_service.dart';
 import 'package:plato.cap/src/enrollments/patron_enrollments.dart';
 
@@ -11,7 +11,7 @@ import 'package:plato.cap/src/enrollments/patron_enrollments.dart';
   selector: 'enrollments',
   providers: [EnrollmentsService]
 )
-class EnrollmentsComponent implements OnInit {
+class EnrollmentsComponent implements AfterViewInit {
   PatronEnrollments get enrollments => _enrollmentsService.enrollments;
 
   final EnrollmentsService _enrollmentsService;
@@ -21,9 +21,9 @@ class EnrollmentsComponent implements OnInit {
   /// The [EnrollmentsComponent] constructor...
   EnrollmentsComponent (this._enrollmentsService);
 
-  /// The [ngOnInit] method...
+  /// The [ngAfterViewInit] method...
   @override
-  void ngOnInit() {}
+  void ngAfterViewInit() {}
 
   /// The [enrollmentSelected] method...
   void enrollmentSelected (PatronEnrollment enrollment) {